[MOB-3544] Update Welcome background video and remove loop
